requirements.txt
matplotlib==3.3.4
speedtest-cli==2.1.2